Create LazyInitialisation.go
Lazy initialization refers to delaying the initialization of a resource until it is needed. In Go, sync.Once is used to ensure that a piece of code runs only once, no matter how many times it is called.

Explanation:
once.Do(loadConfig): The loadConfig function will only be executed once, even though getConfig is called multiple times across different goroutines.
This is useful for initializing global resources, such as configuration, only when they are first accessed.